+++ crates/ruma-client-api/src/media/get_content.rs
@@ -11,7 +11,7 @@ pub mod v3 {
 
     use http::header::{CONTENT_DISPOSITION, CONTENT_TYPE};
     use ruma_common::{
-        IdParseError, MxcUri, OwnedServerName,
+        IdParseError, MxcUri, ServerName,
         api::{auth_scheme::NoAccessToken, request, response},
         http_headers::{CROSS_ORIGIN_RESOURCE_POLICY, ContentDisposition},
         metadata,
@@ -37,7 +37,7 @@ pub mod v3 {
     pub struct Request {
         /// The server name from the mxc:// URI (the authoritory component).
         #[ruma_api(path)]
-        pub server_name: OwnedServerName,
+        pub server_name: ServerName,
 
         /// The media ID from the mxc:// URI (the path component).
         #[ruma_api(path)]
@@ -102,7 +102,7 @@ pub mod v3 {
     #[allow(deprecated)]
     impl Request {
         /// Creates a new `Request` with the given media ID and server name.
-        pub fn new(media_id: String, server_name: OwnedServerName) -> Self {
+        pub fn new(media_id: String, server_name: ServerName) -> Self {
             Self {
                 media_id,
                 server_name,